+++ models/summary.go
@@ -36,7 +36,7 @@ type Summary struct {
 type SummaryItems []*SummaryItem
 
 type SummaryItem struct {
-	ID        uint          `json:"-" gorm:"primary_key"`
+	ID        uint64        `json:"-" gorm:"primary_key"`
 	Summary   *Summary      `json:"-" gorm:"not null; constraint:OnUpdate:CASCADE,OnDelete:CASCADE"`
 	SummaryID uint          `json:"-"`
 	Type      uint8         `json:"-" gorm:"index:idx_type"`
